refactor: split tunnel c2s relay handlers
